Refactoring and formatting.
